+++ site/src/other_tutorial.rs
@@ -13,6 +13,7 @@ pub enum OtherTutorialPage {
     Audio,
     Images,
     Documentation,
+    CodeTactility,
 }
 
 impl OtherTutorialPage {
@@ -30,6 +31,9 @@ impl OtherTutorialPage {
             Self::Audio => Audio().into_view(),
             Self::Images => ImagesAndGifs().into_view(),
             Self::Documentation => Documentation().into_view(),
+            Self::CodeTactility => {
+                title_markdown("Code Tactility", "/text/code_tactility.md", ()).into_view()
+            }
         }
     }
 }
@@ -183,5 +187,14 @@ pub fn Documentation() -> impl IntoView {
         <p>"The "<code>"?"</code>" is similar to the "<Prim prim=Primitive::Stack/>" function because the arguments indicate the intended state of the stack before the function is called."</p>
         <p>"If you also want to give names to a function's outputs, you can list them in front of the "<code>"?"</code>". This lets you read the comment signature right-to-left, the same way as normal Uiua code."</p>
         <Editor example="# Quotient Remainder ? Divisor Dividend\nDivRem ← ⌊⊃÷◿\nDivRem 3 7"/>
+
+        <Hd id="track-caller"><code>"# Track caller!"</code></Hd>
+
+        <p>"When writing code that other people might use, it is important that error messages point to a useful place. When a function errors, the error message often refers to the base built-in function that emitted the error."</p>
+        <Editor example="F ← +@a\nF 5\nF @b"/> // Should fail
+        <p>"However, that position in the code might not be useful to the person trying to figure out where they went wrong."</p>
+        <p>"The "<code>"# Tracker caller!"</code>" semantic comment tells a function to emit errors at its call site rather than from within. It can be place either inside the function's body, or on the line just above it."</p>
+        <Editor example="# Track caller!\nF ← +@a\nF 5\nF @b"/> // Should fail
+        <p>"Notice the difference in the error messages."</p>
     }
 }